+++ app/classes/PaginaCobro.php
@@ -1,10 +1,12 @@
 <?php
 require_once dirname(__FILE__).'/../conf.php';
 
-class PaginaCobro extends Pagina {
-	public function __construct(&$sesion, $index = false) {
-		parent::__construct($sesion, $index);
-	}
+class PaginaCobro extends Pagina
+{
+
+    public function __construct(&$sesion) {
+        parent::__construct($sesion);
+    }
 
 	/***
 	 * PrintPasos
@@ -14,4 +16,5 @@ class PaginaCobro extends Pagina {
 	function PrintPasos($sesion, $paso, $cliente = null, $id_cobro = null, $incluye_gastos = 1, $incluye_honorarios = 1) {
 		require Conf::ServerDir() . '/templates/' . Conf::Templates() . '/top_cobro.php';
 	}
+
 }