+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+Friday 2015, July 24th.
+* **Fix:** [Doblecero en factura PDF](https://github.com/LemontechSA/ttb/pull/1056)
+
+## [Release 15.14.0](https://github.com/LemontechSA/ttb/pull/1055)
+Thursday 2015, July 23th.
+* [Reporte Avanzado corrección trámites](https://github.com/LemontechSA/ttb/pull/1048)
+* [Panama](https://github.com/LemontechSA/ttb/pull/927)
+
+## Hotfix 15.13.1
+Wednesday 2015, July 22th.
+* **Feature:** [Agregar monto al contrato en integración con moreno baldivieso](https://github.com/LemontechSA/ttb/pull/1053)
+
 ## [Release 15.13.0](https://github.com/LemontechSA/ttb/pull/1050)
 Wednesday 2015, July 15th.
 * **Feature:** [Corregir link de tarifas](https://github.com/LemontechSA/ttb/pull/1047)